% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{create_new_memo_to_record}
\alias{create_new_memo_to_record}
\title{Function to create a new Quarto memo to record template file.}
\usage{
create_new_memo_to_record()
}
\value{
A new Quarto memo to record template file. Run this function to create a new Quarto file.
}
\description{
Function to create a new Quarto memo to record template file.
}
