% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_jpe_data}
\alias{calculate_jpe_data}
\title{Calculate JPE Data}
\usage{
calculate_jpe_data(JPI, V.JPI, timing)
}
\arguments{
\item{timing}{Prelminary or final}
}
\value{
A tibble with JPE data
}
\description{
Calculate JPE Data
}
